Fix floaterm#toggle with a name parameter
Currently, the `floaterm#toggle` method doesn't work when a named argument is passed (It will always create a new buffer, even if one has already been created). On inspection, this seems to be an issue with the order of arguments passed to `floaterm#new` - the `jobopts` and `opts` arguments seem to be switched. Reversing them fixes the issue.