Un-silence checktime of floaterm#window#hide
If deleting a file, either from within floaterm itself or while floaterm
is open, the `silent` will silence the 'E211 File "{filename}" no longer
available' error if floaterm is hidden before another call to
`:checktime` is made. A common situation when this can happen is when
using floaterm to check out a branch that doesn't contain all the files
currently being edited.

Given that it's not possible to just add another :checktime to
re-trigger the E211 error (since the original :checktime has updated the
timestamps), it seems appropriate to expect those who don't want the
possible (n)vim output resulting from a :checktime to explicitly silence
suppress it to avoid forcing those who do want the output to resort to
custom workarounds.